refactor(P6): input value에 대한 validation은 soultion에서, 객체에 대한 validation은 객체 팩토리에서 하도록 변경
